FINALLY got pagination to work
Seperated the submitBtn from the thing that loops through the tds and assigns the click events. That was causing the double booking

Seperated by making it so that when you click on the td it stores its rowIndex and colIndex in the modal dialog. Later when submitBtn clicked, it takes that data, passes it into a function that returns the appropriate booking fields needed (e.g startDate and what not).

Also the hover effect was still happening for the tds with the time, so fixed that by giving them a class .time and excluding that from the hover the same way .booked is.